// frontend/jobapptrackerfrontend/src/enums/addjobapptestids/AddJobAppPageTestIds_enum.ts
export enum AddJobAppPageTestIds {
    header = "addJobAppPageHeader",
    jobTitle = "addJobAppPageJobTitle",
    jobTitleError = "addJobAppPageJobTitleHelper",
    companyError = "addJobAppPageCompanyError",
    companyField = "addJobAppPageCompany",
    descriptionField = "addJobAppPageDescription",
    statusField = "addJobAppPageStatus",
    statusFieldError = "addJobAppPageStatusError",
    dateAppliedField = "addJobAppPageDateAppliedField",
    dateAppliedError = "addJobAppPageDateAppliedError",
    jobInterviewCard = "addJobAppPageInterviewCard_",
    submitJobAppButton = "addJobAppPageSubmitJobApp",
    addInterviewButton = "addJobAppPageAddInterview"
}